Tools refactor
