+++ src/api/IBMiContent.ts
@@ -8,7 +8,8 @@ import { AttrOperands, CommandResult, IBMiError, IBMiMember, IBMiObject, IFSFile
 import { ConnectionConfiguration } from './Configuration';
 import { FilterType, parseFilter, singleGenericName } from './Filter';
 import { default as IBMi } from './IBMi';
-import { Tools } from './Tools';
+import * as Tools from './tools';
+import * as VscodeTools from './tools/vscode';
 const tmpFile = util.promisify(tmp.file);
 const readFileAsync = util.promisify(fs.readFile);
 const writeFileAsync = util.promisify(fs.writeFile);
@@ -977,7 +978,7 @@ export default class IBMiContent {
    * Returns MarkdownString HTML content
    */
   objectToToolTip(path: string, object: IBMiObject): string {
-    return Tools.generateTooltipHtmlTable(path, {
+    return VscodeTools.generateTooltipHtmlTable(path, {
       "Type": object.type,
       "Attribute": object.attribute,
       "Text": object.text,
@@ -994,7 +995,7 @@ export default class IBMiContent {
    * Returns MarkdownString HTML content
    */
   async sourcePhysicalFileToToolTip(path: string, object: IBMiObject): Promise<string> {
-    return Tools.generateTooltipHtmlTable(path, {
+    return VscodeTools.generateTooltipHtmlTable(path, {
       "Text": object.text,
       "Members": await this.countMembers(object),
       "Length": object.sourceLength,
@@ -1007,7 +1008,7 @@ export default class IBMiContent {
    * Returns MarkdownString HTML content
    */
   memberToToolTip(path: string, member: IBMiMember): string {
-    return Tools.generateTooltipHtmlTable(path, {
+    return VscodeTools.generateTooltipHtmlTable(path, {
       "Text": member.text,
       "Lines": member.lines,
       "Created": member.created?.toISOString().slice(0, 19).replace(`T`, ` `),
@@ -1019,7 +1020,7 @@ export default class IBMiContent {
    * Returns MarkdownString HTML content
    */
   ifsFileToToolTip(path: string, ifsFile: IFSFile): string {
-    return Tools.generateTooltipHtmlTable(path, {
+    return VscodeTools.generateTooltipHtmlTable(path, {
       "Size": ifsFile.size,
       "Modified": ifsFile.modified ? new Date(ifsFile.modified.getTime() - ifsFile.modified.getTimezoneOffset() * 60 * 1000).toISOString().slice(0, 19).replace(`T`, ` `) : ``,
       "Owner": ifsFile.owner ? ifsFile.owner.toUpperCase() : ``
