+++ src/api/debug/index.ts
@@ -8,7 +8,8 @@ import { instance } from "../../instantiate";
 import { ObjectItem } from "../../typings";
 import { ILELibrarySettings } from "../CompileTools";
 import { ConnectionManager } from "../Configuration";
-import { Tools } from "../Tools";
+import * as Tools from "../tools";
+import * as VscodeTools from "../tools/vscode";
 import { Env, getEnvConfig } from "../local/env";
 import * as certificates from "./certificates";
 import { DEBUG_CONFIG_FILE, DebugConfiguration, getDebugServiceDetails, resetDebugServiceDetails } from "./config";
@@ -257,7 +258,7 @@ export async function initialize(context: ExtensionContext) {
       if (connection) {
         const doSetup = await vscode.window.showWarningMessage(`Do you confirm you want to generate or import a new certificate for the Debug Service?`, { modal: true }, 'Generate', 'Import');
         if (doSetup) {
-          Tools.withContext("code-for-ibmi:debugWorking", async () => {
+          VscodeTools.withContext("code-for-ibmi:debugWorking", async () => {
             if (!(await server.getDebugServiceJob()) || await server.stopService(connection)) {
               const debugConfig = await new DebugConfiguration().load();
               const clearResult = await connection.sendCommand({ command: `rm -f ${debugConfig.getRemoteServiceCertificatePath()} ${debugConfig.getRemoteClientCertificatePath()}` });
@@ -276,7 +277,7 @@ export async function initialize(context: ExtensionContext) {
       }
     }),
     vscode.commands.registerCommand(`code-for-ibmi.debug.setup.remote`, (doSetup?: 'Generate' | 'Import') =>
-      Tools.withContext("code-for-ibmi:debugWorking", async () => {
+      VscodeTools.withContext("code-for-ibmi:debugWorking", async () => {
         const connection = instance.getConnection();
         const content = instance.getContent();
         if (connection && content) {
@@ -337,7 +338,7 @@ export async function initialize(context: ExtensionContext) {
 
     vscode.commands.registerCommand(`code-for-ibmi.debug.setup.local`, () =>
       vscode.window.withProgress({ title: "Downloading Debug Service Certificate", location: vscode.ProgressLocation.Window }, async () =>
-        await Tools.withContext("code-for-ibmi:debugWorking", async () => {
+        await VscodeTools.withContext("code-for-ibmi:debugWorking", async () => {
           const connection = instance.getConnection();
           if (connection) {
             const ptfInstalled = server.debugPTFInstalled();
