+++ src/api/errors/diagnostics.ts
@@ -3,7 +3,8 @@ import * as vscode from "vscode";
 import { FileError } from "../../typings";
 import { GlobalConfiguration } from "../Configuration";
 import Instance from "../Instance";
-import { Tools } from "../Tools";
+import * as Tools from "../tools";
+import * as VscodeTools from "../tools/vscode";
 import { getEvfeventFiles } from "../local/actions";
 import { parseErrors } from "./parser";
 
@@ -175,7 +176,7 @@ export function handleEvfeventLines(lines: string[], instance: Instance, evfeven
           // tabs like we do below.
           if (evfeventInfo.extension) {
             const baseName = file.split(`/`).pop();
-            const openFile = Tools.findExistingDocumentByName(`${baseName}.${evfeventInfo.extension}`);
+            const openFile = VscodeTools.findExistingDocumentByName(`${baseName}.${evfeventInfo.extension}`);
             if (openFile) {
               ileDiagnostics.set(openFile, diagnostics);
               continue;
@@ -185,10 +186,10 @@ export function handleEvfeventLines(lines: string[], instance: Instance, evfeven
       }
 
       if (file.startsWith(`/`)) {
-        ileDiagnostics.set(Tools.findExistingDocumentUri(vscode.Uri.from({ scheme: `streamfile`, path: file })), diagnostics);
+        ileDiagnostics.set(VscodeTools.findExistingDocumentUri(vscode.Uri.from({ scheme: `streamfile`, path: file })), diagnostics);
       }
       else {
-        const memberUri = Tools.findExistingDocumentUri(vscode.Uri.from({ scheme: `member`, path: `/${asp}${file}${evfeventInfo.extension ? `.` + evfeventInfo.extension : ``}` }));
+        const memberUri = VscodeTools.findExistingDocumentUri(vscode.Uri.from({ scheme: `member`, path: `/${asp}${file}${evfeventInfo.extension ? `.` + evfeventInfo.extension : ``}` }));
         ileDiagnostics.set(memberUri, diagnostics);
       }
     }
