+++ src/api/local/git.ts
@@ -1,5 +1,5 @@
 import { ExtensionContext, WorkspaceFolder, commands, window } from "vscode";
-import { Tools } from "../Tools";
+import * as VscodeTools from "../tools/vscode";
 import { getBranchLibraryName } from "./env";
 import { instance } from "../../instantiate";
 import { ConnectionConfiguration, GlobalConfiguration } from "../Configuration";
@@ -9,7 +9,7 @@ import IBMiContent from "../IBMiContent";
 const lastBranch: { [workspaceUri: string]: string } = {};
 
 export function getGitBranch(workspaceFolder: WorkspaceFolder) {
-  const gitApi = Tools.getGitAPI();
+  const gitApi = VscodeTools.getGitAPI();
   if (gitApi) {
     const repo = gitApi.getRepository(workspaceFolder.uri);
     if (repo) {
@@ -19,7 +19,7 @@ export function getGitBranch(workspaceFolder: WorkspaceFolder) {
 }
 
 export function setupGitEventHandler(context: ExtensionContext) {
-  const gitApi = Tools.getGitAPI();
+  const gitApi = VscodeTools.getGitAPI();
 
   if (gitApi) {
     gitApi.onDidOpenRepository((repo) => {
