+++ src/testing/content.ts
@@ -5,7 +5,7 @@ import tmp from 'tmp';
 import util, { TextDecoder } from 'util';
 import { Uri, workspace } from "vscode";
 import { TestSuite } from ".";
-import { Tools } from "../api/Tools";
+import * as Tools from "../api/tools";
 import { getMemberUri } from "../filesystems/qsys/QSysFs";
 import { instance } from "../instantiate";
 import { CommandResult, IBMiObject } from "../typings";
