+++ src/testing/encoding.ts
@@ -3,7 +3,7 @@ import tmp from 'tmp';
 import util, { TextDecoder } from 'util';
 import { Uri, workspace } from "vscode";
 import { TestSuite } from ".";
-import { Tools } from "../api/Tools";
+import * as Tools from "../api/tools";
 import { instance } from "../instantiate";
 import { CommandResult } from "../typings";
 import { getMemberUri } from "../filesystems/qsys/QSysFs";
