+++ src/testing/tools.ts
@@ -1,6 +1,6 @@
 import assert from "assert";
 import { TestSuite } from ".";
-import { Tools } from "../api/Tools";
+import * as Tools from "../api/tools";
 
 export const ToolsSuite: TestSuite = {
   name: `Tools API tests`,
