+++ src/views/debugView.ts
@@ -1,5 +1,6 @@
 import vscode from "vscode";
-import { Tools } from "../api/Tools";
+import * as Tools from "../api/tools";
+import * as VscodeTools from "../api/tools/vscode";
 import { checkClientCertificate, debugKeyFileExists, remoteCertificatesExists, SERVICE_CERTIFICATE } from "../api/debug/certificates";
 import { DebugConfiguration, getDebugServiceDetails } from "../api/debug/config";
 import { DebugJob, getDebugServerJob, getDebugServiceJob, isDebugEngineRunning, readActiveJob, readJVMInfo, startServer, startService, stopServer, stopService } from "../api/debug/server";
@@ -47,9 +48,9 @@ export function initializeDebugBrowser(context: vscode.ExtensionContext) {
     debugTreeViewer,
     vscode.commands.registerCommand("code-for-ibmi.debug.refresh", updateDebugBrowser),
     vscode.commands.registerCommand("code-for-ibmi.debug.refresh.item", (item: DebugItem) => debugBrowser.refresh(item)),
-    vscode.commands.registerCommand("code-for-ibmi.debug.job.start", (item: DebugJobItem) => Tools.withContext(`code-for-ibmi:debugWorking`, () => item.start())),
-    vscode.commands.registerCommand("code-for-ibmi.debug.job.stop", (item: DebugJobItem) => Tools.withContext(`code-for-ibmi:debugWorking`, () => item.stop())),
-    vscode.commands.registerCommand("code-for-ibmi.debug.job.restart", async (item: DebugJobItem) => Tools.withContext(`code-for-ibmi:debugWorking`, async () => await item.stop() && item.start())),
+    vscode.commands.registerCommand("code-for-ibmi.debug.job.start", (item: DebugJobItem) => VscodeTools.withContext(`code-for-ibmi:debugWorking`, () => item.start())),
+    vscode.commands.registerCommand("code-for-ibmi.debug.job.stop", (item: DebugJobItem) => VscodeTools.withContext(`code-for-ibmi:debugWorking`, () => item.stop())),
+    vscode.commands.registerCommand("code-for-ibmi.debug.job.restart", async (item: DebugJobItem) => VscodeTools.withContext(`code-for-ibmi:debugWorking`, async () => await item.stop() && item.start())),
   );
 }
 
@@ -66,7 +67,7 @@ class DebugBrowser implements vscode.TreeDataProvider<BrowserItem> {
   }
 
   async getChildren(item?: DebugItem) {
-    return Tools.withContext(`code-for-ibmi:debugWorking`, async () => item?.getChildren?.() || this.getRootItems());
+    return VscodeTools.withContext(`code-for-ibmi:debugWorking`, async () => item?.getChildren?.() || this.getRootItems());
   }
 
   private async getRootItems() {
