+++ src/views/helpView.ts
@@ -133,16 +133,16 @@ async function downloadLogs() {
       location: vscode.ProgressLocation.Notification,
       title: vscode.l10n.t(`Gathering logs...`),
     }, async () => {
-      const codeForIBMiLog = connection.outputChannelContent;
-      if (codeForIBMiLog !== undefined) {
-        logs.push({
-          label: vscode.l10n.t(`Code for IBM i Log`),
-          detail: `${connection?.currentUser}@${connection?.currentHost}`,
-          picked: true,
-          fileName: 'CodeForIBMi.txt',
-          fileContent: Buffer.from(codeForIBMiLog, 'utf8')
-        });
-      }
+      // const codeForIBMiLog = connection.outputChannelContent;
+      // if (codeForIBMiLog !== undefined) {
+      //   logs.push({
+      //     label: vscode.l10n.t(`Code for IBM i Log`),
+      //     detail: `${connection?.currentUser}@${connection?.currentHost}`,
+      //     picked: true,
+      //     fileName: 'CodeForIBMi.txt',
+      //     fileContent: Buffer.from(codeForIBMiLog, 'utf8')
+      //   });
+      // }
 
       const debugConfig = await new DebugConfiguration().load();
       try {
